requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
Pillow>=8.3.0
opencv-python>=4.5.0
roboflow>=1.1.0
